apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.5.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
slowapi==0.1.9
soundfile==0.12.1
resampy==0.4.3
numpy==1.26.4
scikit-learn==1.5.1
tensorflow==2.17.0
tensorflow-hub==0.16.1
psycopg2-binary==2.9.9
pytest==8.3.2
httpx==0.27.2
